fix(griffin): fix fd leak in tables() function
