+++ core/src/main/java/io/questdb/griffin/engine/functions/catalogue/TableListFunctionFactory.java
@@ -126,6 +126,7 @@ public class TableListFunctionFactory implements FunctionFactory {
             @Override
             public void close() {
                 findPtr = ff.findClose(findPtr);
+                Misc.free(metaReader);//release FD of last table on the list
             }
 
             @Override
