+++ core/src/test/java/io/questdb/cutlass/pgwire/PGJobContextTest.java
@@ -6428,41 +6428,6 @@ create table tab as (
         );
     }
 
-    private static void toSink(InputStream is, CharSink sink) throws IOException {
-        // limit what we print
-        byte[] bb = new byte[1];
-        int i = 0;
-        while (is.read(bb) > 0) {
-            byte b = bb[0];
-            if (i > 0) {
-                if ((i % 16) == 0) {
-                    sink.put('\n');
-                    Numbers.appendHexPadded(sink, i);
-                }
-            } else {
-                Numbers.appendHexPadded(sink, i);
-            }
-            sink.put(' ');
-
-            final int v;
-            if (b < 0) {
-                v = 256 + b;
-            } else {
-                v = b;
-            }
-
-            if (v < 0x10) {
-                sink.put('0');
-                sink.put(hexDigits[b]);
-            } else {
-                sink.put(hexDigits[v / 0x10]);
-                sink.put(hexDigits[v % 0x10]);
-            }
-
-            i++;
-        }
-    }
-
     private void assertHexScript(String script) throws Exception {
         final Rnd rnd = new Rnd();
         assertHexScript(NetworkFacadeImpl.INSTANCE, script, new DefaultPGWireConfiguration() {
@@ -6485,115 +6450,6 @@ create table tab as (
         });
     }
 
-    private void assertResultSet(String expected, StringSink sink, ResultSet rs) throws SQLException, IOException {
-        // dump metadata
-        ResultSetMetaData metaData = rs.getMetaData();
-        final int columnCount = metaData.getColumnCount();
-        for (int i = 0; i < columnCount; i++) {
-            if (i > 0) {
-                sink.put(',');
-            }
-
-            sink.put(metaData.getColumnName(i + 1));
-            sink.put('[').put(JDBCType.valueOf(metaData.getColumnType(i + 1)).name()).put(']');
-        }
-        sink.put('\n');
-
-        while (rs.next()) {
-            for (int i = 1; i <= columnCount; i++) {
-                if (i > 1) {
-                    sink.put(',');
-                }
-                switch (JDBCType.valueOf(metaData.getColumnType(i))) {
-                    case VARCHAR:
-                    case NUMERIC:
-                        String stringValue = rs.getString(i);
-                        if (rs.wasNull()) {
-                            sink.put("null");
-                        } else {
-                            sink.put(stringValue);
-                        }
-                        break;
-                    case INTEGER:
-                        int intValue = rs.getInt(i);
-                        if (rs.wasNull()) {
-                            sink.put("null");
-                        } else {
-                            sink.put(intValue);
-                        }
-                        break;
-                    case DOUBLE:
-                        double doubleValue = rs.getDouble(i);
-                        if (rs.wasNull()) {
-                            sink.put("null");
-                        } else {
-                            sink.put(doubleValue);
-                        }
-                        break;
-                    case TIMESTAMP:
-                        Timestamp timestamp = rs.getTimestamp(i);
-                        if (timestamp == null) {
-                            sink.put("null");
-                        } else {
-                            sink.put(timestamp.toString());
-                        }
-                        break;
-                    case REAL:
-                        float floatValue = rs.getFloat(i);
-                        if (rs.wasNull()) {
-                            sink.put("null");
-                        } else {
-                            sink.put(floatValue, 3);
-                        }
-                        break;
-                    case SMALLINT:
-                        sink.put(rs.getShort(i));
-                        break;
-                    case BIGINT:
-                        long longValue = rs.getLong(i);
-                        if (rs.wasNull()) {
-                            sink.put("null");
-                        } else {
-                            sink.put(longValue);
-                        }
-                        break;
-                    case CHAR:
-                        String strValue = rs.getString(i);
-                        if (rs.wasNull()) {
-                            sink.put("null");
-                        } else {
-                            sink.put(strValue.charAt(0));
-                        }
-                        break;
-                    case BIT:
-                        sink.put(rs.getBoolean(i));
-                        break;
-                    case TIME:
-                    case DATE:
-                        timestamp = rs.getTimestamp(i);
-                        if (rs.wasNull()) {
-                            sink.put("null");
-                        } else {
-                            sink.put(timestamp.toString());
-                        }
-                        break;
-                    case BINARY:
-                        InputStream stream = rs.getBinaryStream(i);
-                        if (rs.wasNull()) {
-                            sink.put("null");
-                        } else {
-                            toSink(stream, sink);
-                        }
-                        break;
-                    default:
-                        assert false;
-                }
-            }
-            sink.put('\n');
-        }
-
-        TestUtils.assertEquals(expected, sink);
-    }
 
     private void assertTransaction(boolean simple) throws Exception {
         assertMemoryLeak(() -> {
